Allow to generate floats using normal distribution
This allow us to generate points with more probability when we are
closer to a limit/key point.

The sum of 6 draws from a uniform distribution give as a draw of a
normal distribution centered in 3. This is commonly used in programming
for simplicity and it is proved by the central limit theorem.